I restructured my code to fulfil more requirements and become easier to read.
Further Description: Implemented a tree data structure to represent the levels and their corresponding heart emojis. Each node in the tree represents a level, and the left and right children represent the subsequent levels. I also separated the code to make the sections more distict for readability purposes. I also removed the unlock_hearts function because it was quite redundant and initialized the level_wins counter